refactor: split plugin into assets-sync library and thin wasip2 wrapper
Business logic (canister types, CanisterCall trait, sync, scan, content)
moves to the new `assets-sync` library crate so unit tests can run on
the host. `plugin` becomes a thin cdylib that only holds the WIT
bindings and the `WasiCanister` production implementation.
