+++ assets-sync/src/canister.rs
@@ -1,0 +1,106 @@
+//! Assets canister API: Candid wire types and call wrappers.
+//!
+//! Types are ported from `ic-asset` (`src/canisters/frontend/ic-asset/src/canister_api/`).
+//! Only the subset needed for the V2 batch-upload flow is included.
+
+#![allow(dead_code)]
+
+use candid::{CandidType, Nat, Principal};
+use serde::Deserialize;
+use std::collections::HashMap;
+
+// --- Candid wire types ---
+
+#[derive(CandidType, Clone, Debug, Deserialize)]
+pub struct AssetEncodingDetails {
+    pub content_encoding: String,
+    pub sha256: Option<Vec<u8>>,
+}
+
+#[derive(CandidType, Clone, Debug, Deserialize)]
+pub struct AssetDetails {
+    pub key: String,
+    pub encodings: Vec<AssetEncodingDetails>,
+    pub content_type: String,
+}
+
+#[derive(CandidType, Clone, Debug)]
+pub struct CreateAssetArguments {
+    pub key: String,
+    pub content_type: String,
+    pub max_age: Option<u64>,
+    pub headers: Option<HashMap<String, String>>,
+    pub enable_aliasing: Option<bool>,
+    pub allow_raw_access: Option<bool>,
+}
+
+#[derive(CandidType, Clone, Debug)]
+pub struct SetAssetContentArguments {
+    pub key: String,
+    pub content_encoding: String,
+    pub chunk_ids: Vec<Nat>,
+    pub last_chunk: Option<Vec<u8>>,
+    pub sha256: Option<Vec<u8>>,
+}
+
+#[derive(CandidType, Clone, Debug)]
+pub struct UnsetAssetContentArguments {
+    pub key: String,
+    pub content_encoding: String,
+}
+
+#[derive(CandidType, Clone, Debug)]
+pub struct DeleteAssetArguments {
+    pub key: String,
+}
+
+#[derive(CandidType, Clone, Debug)]
+pub struct ClearArguments {}
+
+#[derive(CandidType, Clone, Debug)]
+pub struct SetAssetPropertiesArguments {
+    pub key: String,
+    pub max_age: Option<Option<u64>>,
+    pub headers: Option<Option<Vec<(String, String)>>>,
+    pub allow_raw_access: Option<Option<bool>>,
+    pub is_aliased: Option<Option<bool>>,
+}
+
+#[derive(CandidType, Clone, Debug)]
+pub enum BatchOperationKind {
+    Clear(ClearArguments),
+    DeleteAsset(DeleteAssetArguments),
+    CreateAsset(CreateAssetArguments),
+    UnsetAssetContent(UnsetAssetContentArguments),
+    SetAssetContent(SetAssetContentArguments),
+    SetAssetProperties(SetAssetPropertiesArguments),
+}
+
+#[derive(CandidType, Debug)]
+pub struct CommitBatchArguments {
+    pub batch_id: Nat,
+    pub operations: Vec<BatchOperationKind>,
+}
+
+#[derive(CandidType, Clone, Debug, Deserialize)]
+pub enum Permission {
+    Commit,
+    ManagePermissions,
+    Prepare,
+}
+
+// --- Canister call trait ---
+
+pub trait CanisterCall {
+    fn api_version(&self) -> Result<u16, String>;
+    fn list_assets(&self) -> Result<Vec<AssetDetails>, String>;
+    fn create_batch(&self) -> Result<Nat, String>;
+    fn create_chunk(&self, batch_id: &Nat, content: &[u8]) -> Result<Nat, String>;
+    fn commit_batch(&self, args: CommitBatchArguments) -> Result<(), String>;
+    fn list_permitted(&self, permission: Permission) -> Result<Vec<Principal>, String>;
+    fn grant_permission_via_proxy(
+        &self,
+        to_principal: Principal,
+        permission: Permission,
+    ) -> Result<(), String>;
+}
