+++ assets-sync/src/content.rs
