+++ assets-sync/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod canister;
+pub mod content;
+pub mod scan;
+pub mod sync;
