+++ assets-sync/src/scan.rs
