+++ assets-sync/src/sync.rs
