+++ plugin/Cargo.toml
@@ -7,17 +7,8 @@ publish = false
 [lib]
 crate-type = ["cdylib"]
 
-[dev-dependencies]
-tempfile.workspace = true
-
 [dependencies]
-brotli.workspace = true
+assets-sync = { path = "../assets-sync" }
 candid.workspace = true
-flate2.workspace = true
-hex.workspace = true
-mime.workspace = true
-mime_guess.workspace = true
 serde.workspace = true
-serde_bytes.workspace = true
-sha2.workspace = true
 wit-bindgen = { workspace = true, features = ["realloc"] }
